make StorageService into utility functions.
