+++ core/new-gui/src/app/dashboard/component/feature-container/saved-workflow-section/saved-workflow-section.component.ts
@@ -7,7 +7,6 @@ import { Observable } from 'rxjs';
 import { Workflow } from '../../../../common/type/workflow';
 import { Router } from '@angular/router';
 import { WorkflowPersistService } from '../../../../common/service/user/workflow-persist/workflow-persist.service';
-import { StorageService } from '../../../../common/service/storage.service';
 
 /**
  * SavedProjectSectionComponent is the main interface for
@@ -31,8 +30,7 @@ export class SavedWorkflowSectionComponent implements OnInit {
   constructor(
     private workflowPersistService: WorkflowPersistService,
     private modalService: NgbModal,
-    private router: Router,
-    private storageService: StorageService
+    private router: Router
   ) {
   }
